- Thread affinity implementation:    Sleep 10sec. before starting the next NUMA attached process in a data-processing node.
- Run Clara process:
   Coordinate DPE and associated Orchestrator start-up. Start Orchestrator after DPE successful startup only.

- Farm default parameters:
   farm.scale = disabled, farm.memory=0, farm.cpu=0, farm.disk=5GB, farm.time=24hours.

- CLAS12 plugin installation:
   user confirmation request before removing the old installation.

+++ scripts/unix/install-claracre-clas.sh
@@ -9,10 +9,15 @@ if ! [ -n "$CLARA_HOME" ]; then
     exit
 fi
 
+echo "If you have an old installation at $CLARA_HOME it will be deleted."
+read -n 1 -p "Do you want to continue? Y/N" uinput
+if [ "$uinput" = "N" ]; then exit 0
+fi
+
 rm -rf "$CLARA_HOME"
 
 PLUGIN=5a.2.0
-FV=4.3.4
+FV=4.3.5
 
 case "$1" in
     -f | --framework)
